생년월일 - 숫자로 직접입력 - 날짜관련 전부 직접입력 (지불완료날짜, 대리점등록날짜 등) 재류자격 - 직접입력 (현재 배우자비자, 기술인문비자 2가지종류로 선택밖에 없음 ) "월세 관리비 기타비용 쪽에 금액적는 부분에 총금액의 합계는 월세+관리비+기타비용만 들어가야되고 합계금액에 의해 보증수수료계산이 들어가야됩니다 " 입주자 등록화면에 승인번호 입력란 추가 및 입주자 목록에 연동 입주자 등록화면에 승인날짜 입력란 추가 및 입주자 목록에 연동 입주자 등록화면 보증수수료 이름을 보증심사료로 변경 및 입주자목록에 연동, 심사율, 갱신료 연동 입주자 목록  - 승인날짜 기준으로 월별로 구분할수 있는 기능 연체자 목록 - 연체된 월 기준으로 월별로 구분할수 있는 기능 연체자 목록 승인번호 연동 연체자 목록 - 월세를 청구금액으로 변경 및 연동 대리점등록 - 등록날짜 오른쪽에 계약번호 입력란 추가 및 대리점 목록에 연동 대리점 목록에 승인일을 등록일로 이름 변경

+++ src/components/CompanyTable.vue
@@ -1,5 +1,11 @@
 <template>
-  <a-table :columns="columns" :dataSource="companyList" :scroll="windowSize" expandRowByClick @change="handleTableChange">
+  <a-table
+    :columns="columns"
+    :dataSource="companyList"
+    :scroll="windowSize"
+    expandRowByClick
+    @change="handleTableChange"
+  >
     <div slot="expandedRowRender" slot-scope="record" class="detail-row">
       <p style="margin: 0">{{ record.notes }}</p>
       <a-button type="primary" @click="function(){
@@ -21,51 +27,63 @@
 </template>
 <script>
 const columns = [
-  { title: '구분', dataIndex: 'index', key: 'index'},
-  { title: '계약번호', dataIndex: 'approvalNumber', key: 'approvalNumber'},
-  { title: '승인일', dataIndex: 'createdDate', key: 'createdDate'},
-  { title: '대리점 구분', dataIndex: 'companyType', key: 'companyType'},
-  { title: '추심구분', dataIndex: 'debtCollectionType', key: 'debtCollectionType'},
-  { title: '회사명', dataIndex: 'companyName', key: 'companyName'},
-  { title: '회사주소', dataIndex: 'companyAdress', key: 'companyAdress'},
-  { title: '대표자', dataIndex: 'companyOnwer', key: 'companyOnwer'},
-  { title: '대표자TEL', dataIndex: 'companyOnwerTel', key: 'companyOnwerTel'},
-  { title: '시스템관리자', dataIndex: 'systemManager', key: 'systemManager'},
-  { title: '관리자Email', dataIndex: 'systemManagerEmail', key: 'systemManagerEmail'},
-  { title: 'FAX', dataIndex: 'fax', key: 'fax'},
-  { title: '비고', dataIndex: 'notes', key: 'notes'},
-  { title: '등록날짜', dataIndex: 'joinDate', key: 'joinDate'},
-  { title: '보유 물건 수', dataIndex: 'buildingCount', key: 'buildingCount'},
-  { title: '종업원 수', dataIndex: 'employeeCount', key: 'employeeCount'},
-  { title: '상품 종류', dataIndex: 'productType', key: 'productType'},
-  { title: '보증 수수료', dataIndex: 'fee1', key: 'fee1'},
-  { title: '갱신료', dataIndex: 'novationFee', key: 'novationFee'},
-  { title: '대리점 수수료', dataIndex: 'propertyManagermentCompanyFee', key: 'propertyManagermentCompanyFee'},
+  { title: "구분", dataIndex: "index", key: "index" },
+  { title: "계약번호", dataIndex: "approvalNumber", key: "approvalNumber" },
+  { title: "등록일", dataIndex: "createdDate", key: "createdDate" },
+  { title: "대리점 구분", dataIndex: "companyType", key: "companyType" },
+  {
+    title: "추심구분",
+    dataIndex: "debtCollectionType",
+    key: "debtCollectionType"
+  },
+  { title: "회사명", dataIndex: "companyName", key: "companyName" },
+  { title: "회사주소", dataIndex: "companyAdress", key: "companyAdress" },
+  { title: "대표자", dataIndex: "companyOnwer", key: "companyOnwer" },
+  { title: "대표자TEL", dataIndex: "companyOnwerTel", key: "companyOnwerTel" },
+  { title: "시스템관리자", dataIndex: "systemManager", key: "systemManager" },
+  {
+    title: "관리자Email",
+    dataIndex: "systemManagerEmail",
+    key: "systemManagerEmail"
+  },
+  { title: "FAX", dataIndex: "fax", key: "fax" },
+  { title: "비고", dataIndex: "notes", key: "notes" },
+  { title: "등록날짜", dataIndex: "joinDate", key: "joinDate" },
+  { title: "보유 물건 수", dataIndex: "buildingCount", key: "buildingCount" },
+  { title: "종업원 수", dataIndex: "employeeCount", key: "employeeCount" },
+  { title: "상품 종류", dataIndex: "productType", key: "productType" },
+  { title: "보증 심사료", dataIndex: "fee1", key: "fee1" },
+  { title: "갱신료", dataIndex: "novationFee", key: "novationFee" },
+  {
+    title: "대리점 수수료",
+    dataIndex: "propertyManagermentCompanyFee",
+    key: "propertyManagermentCompanyFee"
+  }
 ];
 import { mapGetters } from "vuex";
 import { T } from "../store/module-example/types";
 export default {
   data() {
     return {
-      db:"", // firebase
+      db: "", // firebase
       columns,
-      windowSize:{x:'max-content'}
+      windowSize: { x: "max-content" }
     };
   },
   computed: {
     ...mapGetters({
-      companyList:"getCompanyList",
+      companyList: "getCompanyList"
     })
   },
-  mounted(){
+  mounted() {
     this.db = firebase.firestore();
-    const y = window.innerHeight-300;
+    const y = window.innerHeight - 300;
     this.windowSize = {
-      x:'max-content' 
-    }
+      x: "max-content"
+    };
   },
   methods: {
-    alertMsg({type="info",msg=""}) {
+    alertMsg({ type = "info", msg = "" }) {
       switch (type) {
         case "info":
           this.$message.info(msg);
@@ -80,23 +98,27 @@ export default {
           break;
       }
     },
-    detail(id){
-      this.$store.dispatch(T.CHANGE_UPDATE_COMPNAY_ID,id);
-      this.$store.dispatch(T.CHANGE_TAB_INDEX,20);
-      
+    detail(id) {
+      this.$store.dispatch(T.CHANGE_UPDATE_COMPNAY_ID, id);
+      this.$store.dispatch(T.CHANGE_TAB_INDEX, 20);
     },
-    confirm(id){
+    confirm(id) {
       const thisObj = this;
-      this.db.collection("companys").doc(id).delete().then(function() {
-        thisObj.$store.dispatch(T.DELETE_COMPANY,id);
-        thisObj.alertMsg({type:"success",msg:"삭제 완료"});
-      }).catch(function(error) {
-        console.log(error)
-        thisObj.alertMsg({type:"error",msg:"삭제 실패"});
-      });
+      this.db
+        .collection("companys")
+        .doc(id)
+        .delete()
+        .then(function() {
+          thisObj.$store.dispatch(T.DELETE_COMPANY, id);
+          thisObj.alertMsg({ type: "success", msg: "삭제 완료" });
+        })
+        .catch(function(error) {
+          console.log(error);
+          thisObj.alertMsg({ type: "error", msg: "삭제 실패" });
+        });
     },
-    cancel(){
-      this.alertMsg({type:"error",msg:"취소"})
+    cancel() {
+      this.alertMsg({ type: "error", msg: "취소" });
     },
     handleTableChange(pagination, filters, sorter) {
       console.log(pagination);
@@ -108,24 +130,24 @@ export default {
         page: pagination.current,
         sortField: sorter.field,
         sortOrder: sorter.order,
-        ...filters,
+        ...filters
       });
-    },
+    }
   }
 };
 </script>
 <style lang="scss">
-.ant-table-fixed{
+.ant-table-fixed {
   overflow-x: auto;
-  tr{
-    td{
+  tr {
+    td {
     }
   }
 }
-.detail-row{
+.detail-row {
   display: flex;
   align-items: center;
-  .ant-btn{
+  .ant-btn {
     margin-left: 10px;
   }
 }
